+++ lib.php
@@ -46,6 +46,7 @@ function aiprovider_router_status_checks(): array {
         new \aiprovider_router\check\byokkeys($inspector),
         new \aiprovider_router\check\byokeligibility($inspector),
         new \aiprovider_router\check\budgetrates($inspector),
+        new \aiprovider_router\check\budgethistory($inspector),
     ];
 }